Move code to new file
